make element attributes setter its own function
