+++ src/lib/layouts/Webgl.svelte
@@ -6,6 +6,8 @@
 	import { getContext, onMount, setContext } from 'svelte';
 	import { writable } from 'svelte/store';
 
+	import setElementAttributes from '../utils/setElementAttributes.js';
+
 	/** @type {HTMLCanvasElement} [element] The `<canvas>` tag. Useful for bindings. */
 	export let element = undefined;
 
@@ -21,18 +23,11 @@
 	/** @type {WebGLRenderingContext} [context] The `<canvas>`'s WebGL context. Useful for bindings. */
 	export let context = undefined;
 
-	/** @type {Object|null} [attrs] An object that sets additional attribute values onto the `<canvas>` tag*/
+	/** @type {Object} [attrs] An object of key-value pairs that sets additional attribute values onto the `<canvas>` tag for each key*/
 		export let attrs = null;
 
 	$: if (element && attrs) {
-		const attrKeys = Object.keys(attrs);
-		for (let i = 0; i < attrKeys.length; i++) {
-			const attr = attrKeys[i];
-			const property = attrs[attr];
-			if (!element.hasAttribute(attr) || element.getAttribute(attr) !== property) {
-				element.setAttribute(attr, property);
-			}
-		}
+		setElementAttributes(element, attrs);
 	}
 
 	let testGl;
